ref: move Joi validation to middleware file
