+++ src/app/routes/admin-forms.server.routes.js
@@ -109,57 +109,16 @@ module.exports = function (app) {
    * @returns {SubmissionResponse.model} 400 - submission has bad data and could not be processed
    * @security OTP
    */
-  app.route('/v2/submissions/encrypt/preview/:formId([a-fA-F0-9]{24})').post(
-    celebrate({
-      [Segments.BODY]: Joi.object({
-        responses: Joi.array()
-          .items(
-            Joi.object().keys({
-              _id: Joi.string().required(),
-              answer: Joi.string().allow('').required(),
-              fieldType: Joi.string()
-                .required()
-                .valid(...Object.values(BasicField)),
-              signature: Joi.string().allow(''),
-            }),
-          )
-          .required(),
-        encryptedContent: Joi.string()
-          .custom((value, helpers) => {
-            const parts = String(value).split(/;|:/)
-            if (
-              parts.length !== 3 ||
-              parts[0].length !== 44 || // public key
-              parts[1].length !== 32 || // nonce
-              !parts.every((part) => Joi.string().base64().validate(part))
-            ) {
-              return helpers.error('Invalid encryptedContent.')
-            }
-            return value
-          }, 'encryptedContent')
-          .required(),
-        attachments: Joi.object()
-          .pattern(
-            /^[a-fA-F0-9]{24}$/,
-            Joi.object().keys({
-              encryptedFile: Joi.object().keys({
-                binary: Joi.string().required(),
-                nonce: Joi.string().required(),
-                submissionPublicKey: Joi.string().required(),
-              }),
-            }),
-          )
-          .optional(),
-        isPreview: Joi.boolean().required(),
-        version: Joi.number().required(),
-      }),
-    }),
-    authActiveForm(PermissionLevel.Read),
-    EncryptSubmissionMiddleware.validateAndProcessEncryptSubmission,
-    AdminFormController.passThroughSpcp,
-    VerifiedContentMiddleware.encryptVerifiedSpcpFields,
-    EncryptSubmissionMiddleware.prepareEncryptSubmission,
-    adminForms.passThroughSaveMetadataToDb,
-    SubmissionsMiddleware.sendEmailConfirmations,
-  )
+  app
+    .route('/v2/submissions/encrypt/preview/:formId([a-fA-F0-9]{24})')
+    .post(
+      EncryptSubmissionMiddleware.validateEncryptSubmissionParams,
+      authActiveForm(PermissionLevel.Read),
+      EncryptSubmissionMiddleware.validateAndProcessEncryptSubmission,
+      AdminFormController.passThroughSpcp,
+      VerifiedContentMiddleware.encryptVerifiedSpcpFields,
+      EncryptSubmissionMiddleware.prepareEncryptSubmission,
+      adminForms.passThroughSaveMetadataToDb,
+      SubmissionsMiddleware.sendEmailConfirmations,
+    )
 }
